column tests complete
