--- src/Persistence.DurableTasks/Column/CsvColumnLoadStep.cs
@@ -1,22 +1,0 @@
-﻿using GoodToCode.Persistence.Abstractions;
-using GoodToCode.Persistence.Blob.Csv;
-using System.Collections.Generic;
-using System.IO;
-
-namespace GoodToCode.Persistence.DurableTasks
-{
-    public class CsvColumnLoadStep
-    {
-        private readonly ICsvService service;
-
-        public CsvColumnLoadStep(ICsvService serviceCsv)
-        {
-            service = serviceCsv;
-        }
-
-        public  IEnumerable<ICellData> Execute(Stream CsvStream, int columnToAnalyze)
-        {
-            return service.GetColumn(CsvStream, columnToAnalyze);
-        }
-    }
-}
