--- src/Persistence.DurableTasks/Column/CsvColumnSearchStep.cs
@@ -1,37 +1,0 @@
-﻿using GoodToCode.Persistence.Abstractions;
-using GoodToCode.Persistence.Blob.Csv;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-
-namespace GoodToCode.Persistence.DurableTasks
-{
-    public class CsvColumnSearchStep
-    {
-        private readonly ICsvService service;
-
-        public CsvColumnSearchStep(ICsvService serviceCsv)
-        {
-            service = serviceCsv;
-        }
-
-        public IEnumerable<ICellData> Execute(Stream CsvStream, string documentName, string searchString)
-        {
-            var returnCells = new List<ICellData>();
-            
-            documentName = string.IsNullOrWhiteSpace(documentName) ? $"Analytics-{DateTime.UtcNow:u}" : documentName;
-            var sheet = service.GetSheet(CsvStream);
-            if (!sheet.Rows.Any()) throw new ArgumentException("Passed sheet does not have any rows.");
-            var header = sheet.GetRow(1);
-            var foundCells = header.Cells.Where(c => c.ColumnName.Contains(searchString));
-            foreach(var cell in foundCells)
-            {
-                var newCells = sheet.GetColumn(cell.ColumnIndex);
-                returnCells.AddRange(newCells);
-            }
-
-            return returnCells;
-        }
-    }
-}
