+++ src/Persistence.Tests/Column/Csv_Column_ActivityTests.cs
@@ -22,7 +22,6 @@ namespace GoodToCode.Persistence.Tests
         public IEnumerable<CellEntity> SutRows { get; private set; }
         public Dictionary<string, StringValues> SutReturn { get; private set; }
 
-
         public Csv_Column_Tests()
         {
             logItem = LoggerFactory.CreateLogger<Csv_Column_Tests>();
@@ -37,13 +36,7 @@ namespace GoodToCode.Persistence.Tests
             {
                 var bytes = await FileFactoryService.GetInstance().ReadAllBytesAsync(SutCsvFile);
                 Stream itemToAnalyze = new MemoryStream(bytes);
-
-
-
-
-                var workflow = new CsvColumnLoadStep(new CsvService());
-                var results = workflow.Execute(itemToAnalyze, 3);
-
+                var results = new CsvService().GetColumn(itemToAnalyze, 1);
                 Assert.IsTrue(results.Any(), "No results from analytics service.");
             }
             catch (Exception ex)
@@ -62,9 +55,17 @@ namespace GoodToCode.Persistence.Tests
             {
                 var bytes = await FileFactoryService.GetInstance().ReadAllBytesAsync(SutCsvFile);
                 Stream itemToAnalyze = new MemoryStream(bytes);
-                var workflow = new CsvColumnSearchStep(new CsvService());
-                var results = workflow.Execute(itemToAnalyze, "DocName", "*");
-                Assert.IsTrue(results.Any(), "No results from analytics service.");
+                var returnCells = new List<ICellData>();
+                var sheet = new CsvService().GetSheet(itemToAnalyze);
+                if (!sheet.Rows.Any()) throw new ArgumentException("Passed sheet does not have any rows.");
+                var header = sheet.GetRow(1);
+                var foundCells = header.Cells.Where(c => c.ColumnName.Contains('*'));
+                foreach (var cell in foundCells)
+                {
+                    var newCells = sheet.GetColumn(cell.ColumnIndex);
+                    returnCells.AddRange(newCells);
+                }
+                Assert.IsTrue(returnCells.Any(), "No results from analytics service.");
             }
             catch (Exception ex)
             {
